+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/chat/trigger/ChatTriggerInitializer.java
@@ -1,9 +1,9 @@
 package io.github.sakurawald.fuji.module.initializer.chat.trigger;
 
+import io.github.sakurawald.fuji.core.auxiliary.StringUtil;
 import io.github.sakurawald.fuji.core.document.annotation.ColorBox;
 import io.github.sakurawald.fuji.core.document.annotation.Document;
 import io.github.sakurawald.fuji.core.auxiliary.LogUtil;
-import io.github.sakurawald.fuji.core.auxiliary.StringUtil;
 import io.github.sakurawald.fuji.core.command.executor.CommandExecutor;
 import io.github.sakurawald.fuji.core.command.structure.ExtendedCommandSource;
 import io.github.sakurawald.fuji.core.config.handler.abst.BaseConfigurationHandler;
@@ -12,13 +12,9 @@ import io.github.sakurawald.fuji.module.initializer.ModuleInitializer;
 import io.github.sakurawald.fuji.module.initializer.chat.trigger.config.model.ChatTriggerConfigModel;
 import io.github.sakurawald.fuji.module.initializer.chat.trigger.structure.ChatTrigger;
 import net.minecraft.server.command.ServerCommandSource;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 
 @Document(id = 1751826730890L, value = """
     This module allows you to define magic spells in chat, to execute commands.
@@ -48,8 +44,8 @@ public class ChatTriggerInitializer extends ModuleInitializer {
 
     private static final BaseConfigurationHandler<ChatTriggerConfigModel> config = new ObjectConfigurationHandler<>(BaseConfigurationHandler.CONFIG_JSON, ChatTriggerConfigModel.class);
 
-    @SuppressWarnings("DuplicatedCode")
-    public static String processChatTriggers(ServerCommandSource source, String chatString, CallbackInfo ci) {
+    @SuppressWarnings({"DuplicatedCode", "UnusedReturnValue"})
+    public static String processChatTriggers(ServerCommandSource source, String chatString) {
         /* Log it. */
         LogUtil.debug("Process Chat Triggers: chatString = {}", chatString);
 
@@ -59,15 +55,15 @@ public class ChatTriggerInitializer extends ModuleInitializer {
             .stream()
             .filter(it -> chatString.matches(it.regex))
             .forEach(it -> {
-                /* Replace captured-groups for commands. */
-                Matcher matcher = Pattern
-                    .compile(it.regex)
-                    .matcher(chatString);
+                /* Initialize the matcher. */
+                Matcher matcher = it.getCompiledRegex().matcher(chatString);
                 matcher.find();
+
+                /* Replace captured-groups for commands. */
                 List<String> commands = it.commands
                     .stream()
                     .map(cmd -> StringUtil.substituteGroupPlaceholders(matcher, cmd))
-                    .collect(Collectors.toCollection(ArrayList::new));
+                    .toList();
 
                 /* Execute commands. */
                 LogUtil.debug("Execute commands {} for {}", commands, it);