+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/chat/trigger/config/model/ChatTriggerConfigModel.java
@@ -13,11 +13,11 @@ public class ChatTriggerConfigModel {
         """)
     public List<ChatTrigger> triggers = new ArrayList<>() {
         {
-            this.add(new ChatTrigger("magic", List.of("say magic!")));
-            this.add(new ChatTrigger("i am (.+)", List.of("say You just said: $0", "say Hello $1")));
-            this.add(new ChatTrigger("(?<=^|\\s)item(?=\\s|$)", List.of("run as fake-op %player:name% chat display item")));
-            this.add(new ChatTrigger("(?<=^|\\s)inv(?=\\s|$)", List.of("run as fake-op %player:name% chat display inv")));
-            this.add(new ChatTrigger("(?<=^|\\s)ender(?=\\s|$)", List.of("run as fake-op %player:name% chat display ender")));
+            this.add(ChatTrigger.make("magic", List.of("say magic!")));
+            this.add(ChatTrigger.make("i am (.+)", List.of("say You just said: $0", "say Hello $1")));
+            this.add(ChatTrigger.make("(?<=^|\\s)item(?=\\s|$)", List.of("run as fake-op %player:name% chat display item")));
+            this.add(ChatTrigger.make("(?<=^|\\s)inv(?=\\s|$)", List.of("run as fake-op %player:name% chat display inv")));
+            this.add(ChatTrigger.make("(?<=^|\\s)ender(?=\\s|$)", List.of("run as fake-op %player:name% chat display ender")));
         }
     };
 }