+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
 - [nickname] feature: now the `nickname format` supports the `placeholders`. (#438)
+- [chat.trigger] fix: use vanilla Java Matcher to handle the regex replacement, so it will be consistent for complex regex and replacement.